requirements.txt
torch
torchvision
av
opencv-python-headless
ultralytics
Pillow
